Move lazy components to project/lazy/
