+++ src/components/project/lazy/firebase-auth-own-ui.ts
@@ -1,0 +1,5 @@
+import { createLazyComponent } from "@components/common/util/lazy-component-creater";
+
+export const createLazyAuthUI = () => {
+  return createLazyComponent(() => import('./../firebase-auth-own-ui'), resolved => resolved.FirebaseAuthOwnUI.createComponent({}))
+}
